+++ src/modules/antimixedutf8.c
@@ -621,7 +621,7 @@ int lookalikespam_score(const char *text, char *logbuf, size_t logbufsize)
 	int points = 0;
 	int last_character_was_word_separator = 0;
 	int skip;
-	u_int32_t utfchar;
+	uint32_t utfchar;
 
 	for (p = text; *p; p++)
 	{
@@ -697,7 +697,7 @@ CMD_OVERRIDE_FUNC(override_msg)
 		           log_data_string("scripts", logbuf));
 
 		/* Take the action */
-		retval = take_action(client, cfg.ban_action, cfg.ban_reason, cfg.ban_time, 0);
+		retval = take_action(client, cfg.ban_action, cfg.ban_reason, cfg.ban_time, 0, NULL);
 		if ((retval == BAN_ACT_WARN) || (retval == BAN_ACT_SOFT_WARN))
 		{
 			/* no action */